move functions into a utils file and add tests to clean up
